refactor: address map search review feedback
